requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
attrs==19.2.0
awscli==1.16.258
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
boto3==1.9.253
botocore==1.12.249
cached-property==1.5.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
idna==2.7
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-c-kernel==1.2.2
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.7
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==3.13
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.20.1
rsa==3.4.2
s3transfer==0.2.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
ssm-parameter-store==19.5.0
terminado==0.8.2
testpath==0.4.2
texttable==0.9.1
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.3
uWSGI==2.0.18
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
widgetsnbextension==3.5.1
zipp==0.6.0
